+++ test/notNull_notEmpty_test.dart
@@ -11,6 +11,16 @@ class ObjectWithNotNull extends _$ObjectWithNotNullSerializable {
   var aNotNull;
 }
 
+_fieldAGt123(ObjectWithNotNullIf ownni) => ownni.fieldA != null && ownni.fieldA > 123;
+
+@serializable
+class ObjectWithNotNullIf extends _$ObjectWithNotNullIfSerializable {
+  int fieldA;
+
+  @NotNull(iff: _fieldAGt123, description: "fieldB should not be null if fieldA is greater than 123")
+  int fieldB;
+}
+
 @serializable
 class ObjectWithNotEmpty extends _$ObjectWithNotEmptySerializable {
   @notEmpty
@@ -39,7 +49,31 @@ main() {
       expect(validate(o..aNotNull = 'x').errors, isEmpty);
     });
   });
-  
+
+  group('NotNullIf ->', () {
+
+    var o = new ObjectWithNotNullIf();
+    var expected = {
+      'fieldB': ['fieldB should not be null if fieldA is greater than 123']
+    };
+
+    test('fieldA = null, fieldB = null', () {
+      expect(validate(o).errors, isEmpty);
+    });
+
+    test('fieldA = 11, fieldB = null', () {
+      expect(validate(o..fieldA = 11).errors, isEmpty);
+    });
+
+    test('fieldA = 124, fieldB = null', () {
+      expect(validate(o..fieldA = 124).errors, expected);
+    });
+
+    test('fieldA = 124, fieldB = 1', () {
+      expect(validate(o..fieldA = 124..fieldB = 1).errors, isEmpty);
+    });
+  });
+
   group('NotEmpty ->', () {
 
     var o = new ObjectWithNotEmpty();