+++ test/notNull_notEmpty_test.g.dart
@@ -10,6 +10,7 @@ part of notNull_notEmpty_test;
 _initMirrors() {
   initClassMirrors({
     ObjectWithNotNull: ObjectWithNotNullClassMirror,
+    ObjectWithNotNullIf: ObjectWithNotNullIfClassMirror,
     ObjectWithNotEmpty: ObjectWithNotEmptyClassMirror
   });
   initFunctionMirrors({});
@@ -41,7 +42,7 @@ abstract class _$ObjectWithNotNullSerializable extends SerializableMap {
     throwFieldNotFoundException(key, 'ObjectWithNotNull');
   }
 
-  get keys => ObjectWithNotNullClassMirror.fields.keys;
+  Iterable<String> get keys => ObjectWithNotNullClassMirror.fields.keys;
 }
 
 _ObjectWithNotNull__Constructor(params) => new ObjectWithNotNull();
@@ -59,6 +60,67 @@ const ObjectWithNotNullClassMirror =
   'aNotNull'
 ], setters: const [
   'aNotNull'
+]);
+
+// **************************************************************************
+// Generator: DsonGenerator
+// Target: class ObjectWithNotNullIf
+// **************************************************************************
+
+abstract class _$ObjectWithNotNullIfSerializable extends SerializableMap {
+  int get fieldA;
+  int get fieldB;
+  void set fieldA(int v);
+  void set fieldB(int v);
+
+  operator [](Object key) {
+    switch (key) {
+      case 'fieldA':
+        return fieldA;
+      case 'fieldB':
+        return fieldB;
+    }
+    throwFieldNotFoundException(key, 'ObjectWithNotNullIf');
+  }
+
+  operator []=(Object key, value) {
+    switch (key) {
+      case 'fieldA':
+        fieldA = value;
+        return;
+      case 'fieldB':
+        fieldB = value;
+        return;
+    }
+    throwFieldNotFoundException(key, 'ObjectWithNotNullIf');
+  }
+
+  Iterable<String> get keys => ObjectWithNotNullIfClassMirror.fields.keys;
+}
+
+_ObjectWithNotNullIf__Constructor(params) => new ObjectWithNotNullIf();
+
+const $$ObjectWithNotNullIf_fields_fieldA = const DeclarationMirror(type: int);
+const $$ObjectWithNotNullIf_fields_fieldB =
+    const DeclarationMirror(type: int, annotations: const [
+  const NotNull(
+      description: r'fieldB should not be null if fieldA is greater than 123',
+      iff: _fieldAGt123)
+]);
+
+const ObjectWithNotNullIfClassMirror =
+    const ClassMirror(name: 'ObjectWithNotNullIf', constructors: const {
+  '': const FunctionMirror(
+      parameters: const {}, call: _ObjectWithNotNullIf__Constructor)
+}, fields: const {
+  'fieldA': $$ObjectWithNotNullIf_fields_fieldA,
+  'fieldB': $$ObjectWithNotNullIf_fields_fieldB
+}, getters: const [
+  'fieldA',
+  'fieldB'
+], setters: const [
+  'fieldA',
+  'fieldB'
 ]);
 
 // **************************************************************************
@@ -87,7 +149,7 @@ abstract class _$ObjectWithNotEmptySerializable extends SerializableMap {
     throwFieldNotFoundException(key, 'ObjectWithNotEmpty');
   }
 
-  get keys => ObjectWithNotEmptyClassMirror.fields.keys;
+  Iterable<String> get keys => ObjectWithNotEmptyClassMirror.fields.keys;
 }
 
 _ObjectWithNotEmpty__Constructor(params) => new ObjectWithNotEmpty();